chore: complete repository setup — lint configs, CI, docs generation, and dependency hygiene
Add the infrastructure layer for the repository: golangci-lint and yamlfmt
configs, Renovate for dependency updates, CODEOWNERS, crd-ref-docs templates
for API reference generation, and a restructured CI pipeline with parallel
build/lint/test/e2e stages. Rename the `tools` task group to `install` for
clarity and update all downstream references. Clean up stale design docs and
rejected-config examples that have been superseded by CRD validation and
generated API docs.

- Add golangci-lint config with staticcheck, errcheck, revive, and 14 other linters
- Add yamlfmt config for YAML formatting enforcement
- Add Renovate config with grouped Go/k8s/Actions updates and automerge for patches
- Add CODEOWNERS routing all paths to @datum-cloud/engineering
- Add crd-ref-docs config and Go templates for docs/api/bgp.md generation
- Restructure CI workflow into parallel build → lint/test → e2e stages with concurrency control
- Rename `tools` task group to `install` across Taskfile, AGENTS.md, CLAUDE.md, README, and ARCHITECTURE
- Add `docs:generate` task and wire crd-ref-docs binary installation
- Regenerate docs/api/bgp.md from Go types via crd-ref-docs
- Rewrite NOTICE with normalized dependency entries and license file pointers
- Fix BSD-3-Clause text to actual Go Authors license; add BSD-2-Clause for gocheck
- Delete obsolete docs/design/README.md and docs/examples/rejected-configurations.yaml

+++ api/v1alpha1/peer_types_test.go
@@ -141,7 +141,7 @@ func TestBGPPeerLargePeerASN(t *testing.T) {
 			RouterTarget: RouterTarget{
 				RouterRef: &RouterRef{Name: "test-router"},
 			},
-			PeerASN: int64(maxASN),
+			PeerASN: maxASN,
 			Address: "10.0.0.2",
 			AddressFamilies: []AddressFamily{
 				{AFI: AFIIPv4, SAFI: SAFIUnicast},
@@ -159,7 +159,7 @@ func TestBGPPeerLargePeerASN(t *testing.T) {
 		t.Fatalf("Unmarshal: %v", err)
 	}
 
-	if got.Spec.PeerASN != int64(maxASN) {
+	if got.Spec.PeerASN != maxASN {
 		t.Errorf("PeerASN after round-trip: got %d, want %d", got.Spec.PeerASN, maxASN)
 	}
 }
